# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Property Manager API",
    "description": "Documentation for the Property Manager API",
    "version": "v1"
  },
  "paths": {
    "/PropertyManager/SetPrice/{newPrice}": {
      "put": {
        "tags": [
          "ProperyManager"
        ],
        "summary": "Used to set the new price of the property",
        "description": "This is meant for the initial propery spawn\r\nOne path variable is expected, which is the new price per unit of housing.",
        "operationId": "Set Price per housing unit",
        "parameters": [
          {
            "name": "newPrice",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The new price per unit was set"
          },
          "400": {
            "description": "An error occurred, so the old price per unit was used"
          }
        }
      }
    },
    "/PropertyManager/Property": {
      "put": {
        "tags": [
          "ProperyManager"
        ],
        "summary": "Used to get an available property of a given size",
        "description": "body requires size\r\n\r\n{\r\n\tbody:int\r\n}",
        "operationId": "RequestProperty",
        "responses": {
          "200": {
            "description": "Good"
          },
          "400": {
            "description": "Bad"
          }
        }
      }
    },
    "/PropertyManager/Owner/{propertyID}": {
      "get": {
        "tags": [
          "ProperyManager"
        ],
        "summary": "Used to get the owner ID of a given property",
        "description": "requires the poperty ID in the url",
        "operationId": "GetOwner",
        "parameters": [
          {
            "name": "propertyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Good"
          },
          "400": {
            "description": "Bad"
          }
        }
      }
    },
    "/PropertyManager/Sell": {
      "post": {
        "tags": [
          "ProperyManager"
        ],
        "summary": "Used to list a property on the market to be sold",
        "description": "Body requres ownerID",
        "operationId": "SellProperty",
        "responses": {
          "200": {
            "description": "Good"
          },
          "400": {
            "description": "Bad"
          }
        }
      }
    },
    "/PropertyManager/Rent": {
      "post": {
        "tags": [
          "ProperyManager"
        ],
        "summary": "Used to list a property on the market to be rented",
        "description": "Body requres ownerID",
        "operationId": "RentProperty",
        "responses": {
          "200": {
            "description": "Good"
          },
          "400": {
            "description": "Bad"
          }
        }
      }
    },
    "/PropertyManager/Approval": {
      "put": {
        "tags": [
          "ProperyManager"
        ],
        "summary": "Transfers ownership of property or cancels transfer",
        "description": "Body requires propertyId, sellerId, buyerId, price, approval\r\n\r\n{\r\n    propertyId:long\r\n    sellerId:long\r\n    buyerId:long\r\n    price:string\r\n    approval:bool\r\n}",
        "operationId": "Approval",
        "responses": {
          "200": {
            "description": "Good"
          },
          "400": {
            "description": "Bad"
          }
        }
      }
    }
  },
  "components": { }
}